+++ pkg/inframetadata/internal/hostmap/hostmap_test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"go.opentelemetry.io/collector/pdata/pmetric"
 	conventions "go.opentelemetry.io/collector/semconv/v1.18.0"
 	"go.uber.org/multierr"
 
@@ -16,10 +17,130 @@ import (
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/inframetadata/payload"
 )
 
+func BuildMetric[N int64 | float64](name string, value N) *pmetric.Metric {
+	m := pmetric.NewMetric()
+	m.SetEmptyGauge()
+	m.SetName(name)
+	dp := m.Gauge().DataPoints().AppendEmpty()
+	switch any(value).(type) {
+	case int64:
+		dp.SetIntValue(int64(value))
+	case float64:
+		dp.SetDoubleValue(float64(value))
+	}
+	return &m
+}
+
+func TestStrSliceField(t *testing.T) {
+	tests := []struct {
+		attributes  map[string]any
+		key         string
+		expected    []string
+		expectedOk  bool
+		expectedErr string
+	}{
+		{
+			attributes:  map[string]any{},
+			key:         "nonexistingkey",
+			expected:    nil,
+			expectedOk:  false,
+			expectedErr: "",
+		},
+		{
+			attributes: map[string]any{
+				"host.ip": "192.168.1.1",
+			},
+			key:         "host.ip",
+			expected:    nil,
+			expectedOk:  false,
+			expectedErr: "\"host.ip\" has type \"Str\", expected type \"Slice\" instead",
+		},
+		{
+			attributes: map[string]any{
+				"host.ip": []any{},
+			},
+			key:         "host.ip",
+			expected:    nil,
+			expectedOk:  false,
+			expectedErr: "\"host.ip\" is an empty slice, expected at least one item",
+		},
+		{
+			attributes: map[string]any{
+				"host.ip": []any{"192.168.1.1", true},
+			},
+			key:         "host.ip",
+			expected:    nil,
+			expectedOk:  false,
+			expectedErr: "host.ip[1] has type \"Bool\", expected type \"Str\" instead",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.key+"/"+tt.expectedErr, func(t *testing.T) {
+			res := testutils.NewResourceFromMap(t, tt.attributes)
+			actual, ok, err := strSliceField(res.Attributes(), tt.key)
+			assert.Equal(t, tt.expected, actual)
+			assert.Equal(t, tt.expectedOk, ok)
+			if tt.expectedErr != "" {
+				assert.EqualError(t, err, tt.expectedErr)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestIsIPv4(t *testing.T) {
+	// Test cases come from https://stackoverflow.com/a/48519490
+	tests := []struct {
+		ip     string
+		isIPv4 bool
+	}{
+		{ip: "192.168.0.1", isIPv4: true},
+		{ip: "192.168.0.1:80", isIPv4: true},
+		{ip: "::FFFF:C0A8:1", isIPv4: false},
+		{ip: "::FFFF:C0A8:0001", isIPv4: false},
+		{ip: "0000:0000:0000:0000:0000:FFFF:C0A8:1", isIPv4: false},
+		{ip: "::FFFF:C0A8:1%1", isIPv4: false},
+		{ip: "::FFFF:192.168.0.1", isIPv4: false},
+		{ip: "[::FFFF:C0A8:1]:80", isIPv4: false},
+		{ip: "[::FFFF:C0A8:1%1]:80", isIPv4: false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.ip, func(t *testing.T) {
+			assert.Equal(t, tt.isIPv4, isIPv4(tt.ip))
+		})
+	}
+}
+
+func TestIEEERAToGolangFormat(t *testing.T) {
+	tests := []struct {
+		ieeeRA       string
+		golangFormat string
+	}{
+		{
+			ieeeRA:       "AB-01-00-00-00-00-00-00",
+			golangFormat: "ab:01:00:00:00:00:00:00",
+		},
+		{
+			ieeeRA:       "AB-CD-EF-00-00-00",
+			golangFormat: "ab:cd:ef:00:00:00",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.ieeeRA, func(t *testing.T) {
+			assert.Equal(t, tt.golangFormat, ieeeRAtoGolangFormat(tt.ieeeRA))
+		})
+	}
+}
+
 func TestUpdate(t *testing.T) {
 	hostInfo := []struct {
 		hostname        string
 		attributes      map[string]any
+		metric          *pmetric.Metric
 		expectedChanged bool
 		expectedErrs    []string
 	}{
@@ -30,6 +151,7 @@ func TestUpdate(t *testing.T) {
 				conventions.AttributeHostID:        "host-1-hostid",
 				conventions.AttributeHostName:      "host-1-hostname",
 				conventions.AttributeOSDescription: "Fedora Linux",
+				conventions.AttributeOSType:        conventions.AttributeOSTypeLinux,
 				conventions.AttributeHostArch:      conventions.AttributeHostArchAMD64,
 				attributeKernelName:                "GNU/Linux",
 				attributeKernelRelease:             "5.19.0-43-generic",
@@ -40,7 +162,10 @@ func TestUpdate(t *testing.T) {
 				attributeHostCPUModelName:          "11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz",
 				attributeHostCPUStepping:           1,
 				attributeHostCPUCacheL2Size:        12288000,
+				attributeHostIP:                    []any{"192.168.1.140", "fe80::abc2:4a28:737a:609e"},
+				attributeHostMAC:                   []any{"AC-DE-48-23-45-67", "AC-DE-48-23-45-67-01-9F"},
 			},
+			metric:          BuildMetric[int64](metricSystemCPUPhysicalCount, 32),
 			expectedChanged: false,
 		},
 		{
@@ -52,6 +177,7 @@ func TestUpdate(t *testing.T) {
 				conventions.AttributeHostName:      "host-1-hostname",
 				conventions.AttributeOSDescription: "Fedora Linux",
 			},
+			metric:          BuildMetric[float64](metricSystemCPUFrequency, 400_000_005.5),
 			expectedChanged: false,
 		},
 		{
@@ -113,6 +239,9 @@ func TestUpdate(t *testing.T) {
 		} else {
 			assert.NoError(t, err)
 		}
+		if info.metric != nil {
+			hostMap.UpdateFromMetric(info.hostname, *info.metric)
+		}
 	}
 
 	hosts := hostMap.Flush()
@@ -134,6 +263,8 @@ func TestUpdate(t *testing.T) {
 			fieldPlatformProcessor:        "amd64",
 			fieldPlatformMachine:          "amd64",
 			fieldPlatformHardwarePlatform: "amd64",
+			fieldPlatformGOOS:             "linux",
+			fieldPlatformGOOARCH:          "amd64",
 			fieldPlatformKernelName:       "GNU/Linux",
 			fieldPlatformKernelRelease:    "5.19.0-43-generic",
 			fieldPlatformKernelVersion:    "#82~18.04.1-Ubuntu SMP Fri Apr 16 15:10:02 UTC 2021",
@@ -145,10 +276,16 @@ func TestUpdate(t *testing.T) {
 			fieldCPUModelName: "11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz",
 			fieldCPUStepping:  "1",
 			fieldCPUVendorID:  "GenuineIntel",
+			fieldCPUCores:     "32",
+			fieldCPUMHz:       "400.0000055",
+		})
+		assert.Equal(t, md.Payload.Gohai.Gohai.Network, map[string]string{
+			fieldNetworkIPAddressIPv4: "192.168.1.140",
+			fieldNetworkIPAddressIPv6: "fe80::abc2:4a28:737a:609e",
+			fieldNetworkMACAddress:    "ac:de:48:23:45:67",
 		})
 		assert.Nil(t, md.Payload.Gohai.Gohai.FileSystem)
 		assert.Nil(t, md.Payload.Gohai.Gohai.Memory)
-		assert.Nil(t, md.Payload.Gohai.Gohai.Network)
 	}
 
 	if assert.Contains(t, hosts, "host-2-hostid") {
@@ -164,11 +301,12 @@ func TestUpdate(t *testing.T) {
 			fieldPlatformProcessor:        "arm64",
 			fieldPlatformMachine:          "arm64",
 			fieldPlatformHardwarePlatform: "arm64",
+			fieldPlatformGOOARCH:          "arm64",
 		})
 		assert.Empty(t, md.Payload.Gohai.Gohai.CPU)
+		assert.Empty(t, md.Payload.Gohai.Gohai.Network)
 		assert.Nil(t, md.Payload.Gohai.Gohai.FileSystem)
 		assert.Nil(t, md.Payload.Gohai.Gohai.Memory)
-		assert.Nil(t, md.Payload.Gohai.Gohai.Network)
 	}
 
 	assert.Empty(t, hostMap.Flush(), "returned map must be empty after double flush")