+++ pkg/otlp/logs/transform.go
@@ -19,7 +19,6 @@ import (
 	"encoding/hex"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/DataDog/datadog-api-client-go/v2/api/datadog"
 	"github.com/DataDog/datadog-api-client-go/v2/api/datadogV2"
@@ -62,9 +61,9 @@ const (
 
 // Transform converts the log record in lr, which came in with the resource in res to a Datadog log item.
 // the variable specifies if the log body should be sent as an attribute or as a plain message.
+// Deprecated: use Translator instead.
 func Transform(lr plog.LogRecord, res pcommon.Resource, logger *zap.Logger) datadogV2.HTTPLogItem {
 	host, service := extractHostNameAndServiceName(res.Attributes(), lr.Attributes())
-
 	l := datadogV2.HTTPLogItem{
 		AdditionalProperties: make(map[string]string),
 	}
@@ -85,7 +84,7 @@ func Transform(lr plog.LogRecord, res pcommon.Resource, logger *zap.Logger) data
 			l.Message = v.AsString()
 		case "status", "severity", "level", "syslog.severity":
 			status = v.AsString()
-		case "traceid", "contextmap.traceid", "oteltraceid":
+		case "traceid", "trace_id", "contextmap.traceid", "oteltraceid":
 			traceID, err := decodeTraceID(v.AsString())
 			if err != nil {
 				logger.Warn("failed to decode trace id",
@@ -97,7 +96,7 @@ func Transform(lr plog.LogRecord, res pcommon.Resource, logger *zap.Logger) data
 				l.AdditionalProperties[ddTraceID] = strconv.FormatUint(traceIDToUint64(traceID), 10)
 				l.AdditionalProperties[otelTraceID] = v.AsString()
 			}
-		case "spanid", "contextmap.spanid", "otelspanid":
+		case "spanid", "span_id", "contextmap.spanid", "otelspanid":
 			spanID, err := decodeSpanID(v.AsString())
 			if err != nil {
 				logger.Warn("failed to decode span id",
@@ -114,7 +113,10 @@ func Transform(lr plog.LogRecord, res pcommon.Resource, logger *zap.Logger) data
 			tagStr := strings.Join(tags, ",")
 			l.Ddtags = datadog.PtrString(tagStr)
 		default:
-			l.AdditionalProperties[k] = v.AsString()
+			m := flattenAttribute(k, v, 1)
+			for k, v := range m {
+				l.AdditionalProperties[k] = v
+			}
 		}
 		return true
 	})
@@ -156,7 +158,7 @@ func Transform(lr plog.LogRecord, res pcommon.Resource, logger *zap.Logger) data
 	if lr.Timestamp() != 0 {
 		// we are retaining the nano second precision in this property
 		l.AdditionalProperties[otelTimestamp] = strconv.FormatInt(lr.Timestamp().AsTime().UnixNano(), 10)
-		l.AdditionalProperties[ddTimestamp] = lr.Timestamp().AsTime().Format(time.RFC3339)
+		l.AdditionalProperties[ddTimestamp] = lr.Timestamp().AsTime().Format("2006-01-02T15:04:05.000Z07:00")
 	}
 	if l.Message == "" {
 		// set the Message to the Body in case it wasn't already parsed as part of the attributes
@@ -170,6 +172,26 @@ func Transform(lr plog.LogRecord, res pcommon.Resource, logger *zap.Logger) data
 	}
 
 	return l
+}
+
+func flattenAttribute(key string, val pcommon.Value, depth int) map[string]string {
+	result := make(map[string]string)
+
+	if val.Type() != pcommon.ValueTypeMap || depth == 10 {
+		result[key] = val.AsString()
+		return result
+	}
+
+	val.Map().Range(func(k string, v pcommon.Value) bool {
+		newKey := key + "." + k
+		nestedResult := flattenAttribute(newKey, v, depth+1)
+		for nk, nv := range nestedResult {
+			result[nk] = nv
+		}
+		return true
+	})
+
+	return result
 }
 
 func extractHostNameAndServiceName(resourceAttrs pcommon.Map, logAttrs pcommon.Map) (host string, service string) {