+++ pkg/otlp/metrics/metrics_translator.go
@@ -23,21 +23,31 @@ import (
 	"strings"
 	"time"
 
+	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes"
+	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes/source"
+	"github.com/DataDog/opentelemetry-mapping-go/pkg/quantile"
+	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/pmetric"
+	"go.opentelemetry.io/otel/attribute"
+	otelmetric "go.opentelemetry.io/otel/metric"
 	"go.uber.org/zap"
 	"golang.org/x/exp/slices"
 
-	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes"
-	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes/source"
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/metrics/internal/instrumentationlibrary"
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/metrics/internal/instrumentationscope"
-	"github.com/DataDog/opentelemetry-mapping-go/pkg/quantile"
 )
 
 const (
 	metricName             string = "metric name"
 	errNoBucketsNoSumCount string = "no buckets mode and no send count sum are incompatible"
+
+	meterName               string = "github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes"
+	missingSourceMetricName string = "datadog.otlp_translator.resources.missing_source"
+)
+
+var (
+	signalTypeSet = attribute.NewSet(attribute.String("signal", "metrics"))
 )
 
 var _ source.Provider = (*noSourceProvider)(nil)
@@ -48,11 +58,16 @@ func (*noSourceProvider) Source(context.Context) (source.Source, error) {
 	return source.Source{Kind: source.HostnameKind, Identifier: ""}, nil
 }
 
+type instruments struct {
+	missingSources otelmetric.Int64Counter
+}
+
 // Translator is a metrics translator.
 type Translator struct {
-	prevPts *ttlCache
-	logger  *zap.Logger
-	cfg     translatorConfig
+	prevPts     *ttlCache
+	logger      *zap.Logger
+	instruments instruments
+	cfg         translatorConfig
 }
 
 // Metadata specifies information about the outcome of the MapMetrics call.
@@ -62,14 +77,13 @@ type Metadata struct {
 }
 
 // NewTranslator creates a new translator with given options.
-func NewTranslator(logger *zap.Logger, options ...TranslatorOption) (*Translator, error) {
+func NewTranslator(set component.TelemetrySettings, options ...TranslatorOption) (*Translator, error) {
 	cfg := translatorConfig{
 		HistMode:                             HistogramModeDistributions,
 		SendHistogramAggregations:            false,
 		Quantiles:                            false,
 		NumberMode:                           NumberModeCumulativeToDelta,
 		InitialCumulMonoValueMode:            InitialCumulMonoValueModeAuto,
-		ResourceAttributesAsTags:             false,
 		InstrumentationLibraryMetadataAsTags: false,
 		sweepInterval:                        1800,
 		deltaTTL:                             3600,
@@ -89,10 +103,21 @@ func NewTranslator(logger *zap.Logger, options ...TranslatorOption) (*Translator
 	}
 
 	cache := newTTLCache(cfg.sweepInterval, cfg.deltaTTL)
+	meter := set.MeterProvider.Meter(meterName)
+	missingSources, err := meter.Int64Counter(
+		missingSourceMetricName,
+		otelmetric.WithDescription("OTLP resources that are missing a source (e.g. hostname)"),
+		otelmetric.WithUnit("{resource}"),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("failed to build missing source counter: %w", err)
+	}
+
 	return &Translator{
-		prevPts: cache,
-		logger:  logger.With(zap.String("component", "metrics translator")),
-		cfg:     cfg,
+		prevPts:     cache,
+		logger:      set.Logger.With(zap.String("component", "metrics translator")),
+		instruments: instruments{missingSources: missingSources},
+		cfg:         cfg,
 	}, nil
 }
 
@@ -525,16 +550,16 @@ func (t *Translator) mapSummaryMetrics(
 	}
 }
 
-func (t *Translator) source(m pcommon.Map) (source.Source, error) {
-	src, ok := attributes.SourceFromAttrs(m)
-	if !ok {
+func (t *Translator) source(ctx context.Context, m pcommon.Map) (source.Source, bool, error) {
+	src, hasSource := attributes.SourceFromAttrs(m)
+	if !hasSource {
 		var err error
-		src, err = t.cfg.fallbackSourceProvider.Source(context.Background())
+		src, err = t.cfg.fallbackSourceProvider.Source(ctx)
 		if err != nil {
-			return source.Source{}, fmt.Errorf("failed to get fallback source: %w", err)
+			return source.Source{}, false, fmt.Errorf("failed to get fallback source: %w", err)
 		}
 	}
-	return src, nil
+	return src, hasSource, nil
 }
 
 // extractLanguageTag appends a new language tag to languageTags if a new language tag is found from the given name
@@ -655,10 +680,14 @@ func (t *Translator) MapMetrics(ctx context.Context, md pmetric.Metrics, consume
 			consumer.ConsumeAPMStats(sp)
 			continue
 		}
-		src, err := t.source(rm.Resource().Attributes())
+		src, hasSource, err := t.source(ctx, rm.Resource().Attributes())
 		if err != nil {
 			return metadata, err
 		}
+		if !hasSource {
+			t.instruments.missingSources.Add(ctx, 1, otelmetric.WithAttributeSet(signalTypeSet))
+		}
+
 		var host string
 		switch src.Kind {
 		case source.HostnameKind:
@@ -694,6 +723,16 @@ func (t *Translator) MapMetrics(ctx context.Context, md pmetric.Metrics, consume
 			newMetrics := pmetric.NewMetricSlice()
 			for k := 0; k < metricsArray.Len(); k++ {
 				md := metricsArray.At(k)
+				if md.Name() == keyStatsPayload && md.Type() == pmetric.MetricTypeSum {
+
+					// these metrics are an APM Stats payload; consume it as such
+					for l := 0; l < md.Sum().DataPoints().Len(); l++ {
+						if payload, ok := md.Sum().DataPoints().At(l).Attributes().Get(keyStatsPayload); ok && t.cfg.statsOut != nil && payload.Type() == pcommon.ValueTypeBytes {
+							t.cfg.statsOut <- payload.Bytes().AsRaw()
+						}
+					}
+					continue
+				}
 				if v, ok := runtimeMetricsMappings[md.Name()]; ok {
 					metadata.Languages = extractLanguageTag(md.Name(), metadata.Languages)
 					for _, mp := range v {